+++ crates/affinidi-messaging/affinidi-messaging-mediator/tests/integration_test.rs
@@ -70,13 +70,13 @@ async fn test_mediator_server() {
         .await
         .unwrap();
     let alice_secrets_resolver = SimpleSecretsResolver::new(&[
-        Secret::from_str(&format!("{}#key-1", ALICE_DID), &ALICE_V1),
-        Secret::from_str(&format!("{}#key-2", ALICE_DID), &ALICE_E1),
+        Secret::from_str(&format!("{}#key-1", ALICE_DID), &ALICE_V1).expect("Couldn't create Key"),
+        Secret::from_str(&format!("{}#key-2", ALICE_DID), &ALICE_E1).expect("Couldn't create Key"),
     ])
     .await;
     let bob_secrets_resolver = SimpleSecretsResolver::new(&[
-        Secret::from_str(&format!("{}#key-1", BOB_DID), &BOB_V1),
-        Secret::from_str(&format!("{}#key-2", BOB_DID), &BOB_E1),
+        Secret::from_str(&format!("{}#key-1", BOB_DID), &BOB_V1).expect("Couldn't create Key"),
+        Secret::from_str(&format!("{}#key-2", BOB_DID), &BOB_E1).expect("Couldn't create Key"),
     ])
     .await;