+++ crates/affinidi-messaging/CHANGELOG.md
@@ -8,6 +8,31 @@ we find little issues that only affect deployment.
 Missing versions on the changelog simply reflect minor deployment changes on our
 tooling.
 
+## 29th May 2025
+
+### DIDComm Library (0.10.8)
+
+* **FIX:** Clarified error message when `unpack/authcrypt()` fails to deserialize
+a payload successfully
+* **MAINTENANCE:** Crate dependencies updated
+
+### Helpers (0.10.8)
+
+* **MAINTENANCE:** Crate dependencies updated (crate SSI)
+
+### Mediator (0.10.9)
+
+* **MAINTENANCE:** Crate dependencies updated (crate SSI)
+  * **MAINTENANCE:** Includes `affinidi-messaging-meditor-comoon` and `affinidi-messaging-mediator-processors`
+
+### SDK (0.11.2)
+
+* **MAINTENANCE:** Crate dependencies updated (crate SSI)
+
+### Text-Client (0.10.8)
+
+* **MAINTENANCE:** Crate dependencies updated
+
 ## 9th May 2025
 
 ### Mediator (0.10.8)